requirements/base.txt
psycopg2-binary~=2.8
mock~=3.0
python-dotenv
